app/requirements.txt
Flask==2.3.2
pymongo==4.6.3
azure-storage-blob==12.17.0
azure-identity==1.11.0
azure-keyvault-secrets==4.8.0